requirements.txt
# Core Web Framework
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
flask-login>=0.6.0

# Database Drivers
SQLAlchemy>=2.0.0
mysql-connector-python>=8.0.0
psycopg2-binary>=2.9.0

# Environment and Configuration
python-dotenv>=1.0.0

# Web Server
gunicorn>=20.1.0

# LTL Grammar and Parsing
antlr4-python3-runtime==4.13.1

# User Agent Processing
user-agents>=2.2.0
ua-parser>=0.10.0

# HTTP and Networking
requests>=2.31.0
urllib3>=2.0.0
certifi>=2024.07.04

# Utility Libraries
blinker>=1.6.0
click>=8.0.0
charset-normalizer>=3.0.0
idna>=3.4

# Async Support
anyio>=4.0.0
h11>=0.14.0
httpcore>=1.0.0
httpx>=0.25.0
sniffio>=1.3.0

# Template Engine and Security
Jinja2>=3.1.0
MarkupSafe>=2.1.0
itsdangerous>=2.1.0
Werkzeug>=2.3.0

# Data Processing and Validation
pydantic>=2.5.0
pydantic_core>=2.14.0
typing_extensions>=4.8.0

# Parsing and Logic
lark>=1.1.0
pylogics>=0.2.0

# Development and Compatibility
importlib-metadata>=6.0.0
zipp>=3.15.0
greenlet>=2.0.0
exceptiongroup>=1.1.0
distro>=1.8.0
annotated-types>=0.5.0

# Development and Testing
pytest>=7.0.0
pyyaml>=6.0.0

# Image Processing
Pillow>=10.0.0

# Optional packages (comment out if not needed)
# cached_method>=0.1.0
# networkx>=3.0.0
# openai>=1.10.0